treat bare global reads as dead code
Technically this is unsound because it can throw an exception or run a global accessor property. But if you rely on this (instead of, e.g., using typeof), you're on your own.

(Making this change because SWC simplify transform also does this, so it'll happen anyways, and it's better to do it early.)