redundant_obj: invalidate based on props instead of objects to correctly handle aliasing
This allows us to better optimize escaping objects, but pessimizes optimizations for objects that do not alias.
That's okay because this pass is mainly for optimizing reads/writes to unknown objects, e.g.
```
function(a, b, c) {
  a.foo = 1;
  var x = a.foo;
  var y = b.foo;
}
```
(where we could never tell if `a` and `b` alias without interprocedural alias analysis)

In situations where an object clearly does not alias, the sroa pass will almost always take care of it,
so it doesn't really matter if this pass handles it or not.